# Hoist deciding which asyncio current_task to use to module definition time.