@kezabelle 2021-08-16
Additional note: looks like the same hoisting could/should be done for `asgiref.timeout.current_task` but I only just spotted that, so I'll avoid doing it unless the principle behind the PR as it stands is acceptable.