@andrewgodwin 2021-08-16
Yup, this is totally sensible - thanks. Doing it for current_task would also be good.